Reorganize daos.
